+++ src/routes/account/index.jsx
@@ -3,28 +3,11 @@ import * as styles from "./index.module.scss";
 
 import ListArticles from "/src/components/others/listArticles";
 import { getArticlesByAuthor } from "/src/data/ethereumProvider";
-import { ipfsGateway } from "/src/utils/addToIPFS";
+import populateArticleContents from "../../utils/populateArticleContents";
 
 export async function loader({ params }) {
   const articles = await getArticlesByAuthor(params.chain, params.id);
-  const fetchPromises = articles.map(async (article) => {
-    if (!article) return null;
-    try {
-      const response = await fetch(ipfsGateway + article?.articleID);
-      if (!response.ok) {
-        throw new Error("Network response was not OK");
-      }
-      const { title, description } = await response.json();
-      article.title = title;
-      article.description = description;
-    } catch (error) {
-      console.error(error);
-      throw new Error(error.message);
-    }
-  });
-
-  await Promise.all(fetchPromises);
-  return articles;
+  return await populateArticleContents(articles);
 }
 
 export default function Account() {
