+++ src/routes/browse/index.jsx
@@ -7,28 +7,11 @@ import ListArticles from "/src/components/others/listArticles";
 import SyncStatus from "/src/components/presentational/syncStatus";
 
 import { EthereumContext, networkMap, getAllArticles } from "/src/data/ethereumProvider";
-import { ipfsGateway } from "/src/utils/addToIPFS";
+import populateArticleContents from "../../utils/populateArticleContents";
 
 export async function loader({ params }) {
   const articles = await getAllArticles(params.chain);
-  const fetchPromises = articles.map(async (article) => {
-    if (!article) return null;
-    try {
-      const response = await fetch(ipfsGateway + article?.articleID);
-      if (!response.ok) {
-        throw new Error("Network response was not OK");
-      }
-      const { title, description } = await response.json();
-      article.title = title;
-      article.description = description;
-    } catch (error) {
-      console.error(error);
-      throw new Error(error.message);
-    }
-  });
-
-  await Promise.all(fetchPromises);
-  return articles;
+  return await populateArticleContents(articles);
 }
 
 export default function Browse() {
